Fix tranpose-words for multiple selections
The keys executed are the same -- except for an omission of the search
pattern for the second ‘<a-?>’, since the same pattern is used
automatically now that we are using just one execute-keys command. The
only real difference is that now we’re passing ‘-itersel’ to
prevent issues when using transpose-words with multiple cursors.